+++ cmd/fromXsv.go
@@ -22,23 +22,42 @@ THE SOFTWARE.
 package cmd
 
 import (
-	"fmt"
+	"log/slog"
+	"os"
+	"path/filepath"
 
+	"github.com/sfborg/sf/internal/io/xsv"
+	"github.com/sfborg/sf/pkg/config"
 	"github.com/spf13/cobra"
 )
 
 // fromXsvCmd represents the fromXsv command
 var fromXsvCmd = &cobra.Command{
-	Use:   "xsv",
+	Use:   "xsv src-file-or-url output",
 	Short: "Converts CSV/TSV/PSV fiels to SFGA format",
-	Long: `A longer description that spans multiple lines and likely contains examples
-and usage of using your command. For example:
-
-Cobra is a CLI library for Go that empowers applications.
-This application is a tool to generate the needed files
-to quickly create a Cobra application.`,
+	Long: `Takes a Comma/Tab/Pipe Separated Value (XSV) file and converts it to
+a Species File Group Archive file. The file must have headers where at least
+some DarwinCore or CoLDP terms are provided. The "ScientificName" field is
+required. The resuting SFGA file can later be used to compare it with other
+SFGA files, or be converted to other common formats.
+`,
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("fromXsv called")
+		if len(args) != 2 {
+			cmd.Help()
+			os.Exit(0)
+		}
+		src := args[0]
+		out := args[1]
+
+		cfg := config.New()
+		xsv := xsv.New(cfg)
+
+		err := xsv.Import(src, out)
+		if err != nil {
+			file := filepath.Base(src)
+			slog.Error("Cannot import XSV file", "file", file, "error", err)
+			os.Exit(1)
+		}
 	},
 }
 
@@ -53,5 +72,8 @@ func init() {
 
 	// Cobra supports local flags which will only run when this command
 	// is called directly, e.g.:
-	// fromXsvCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+	fromXsvCmd.Flags().StringP("delimiter", "d", "",
+		`delimiter used in the file:
+  auto, csv, tsv, psv are supported
+  auto is default.`)
 }